requirements.txt
streamlit
opencv-python
torch
torchvision
numpy
pillow
scikit-image
mtcnn
reportlab
timm